Split CLI implementation into a few separate files
