+++ src/lib.rs
@@ -1,5 +1,7 @@
 mod app;
 mod data;
+mod cli;
 
 pub use app::{AppState, CueballApp, Project};
-pub use data::{CueList, Cue, CueRunnable, CueTimed, RemarkCue};
+pub use data::{CueList, Cue, CueRunnable, CueTimed, RemarkCue, CLIMode};
+pub use cli::cueball_cli;
